[test] Add a 32-node graph and test the genetic algorithm with specific parameters
Added a medium-sized graph with 32 nodes. I tested the genetic algorithm using the following parameters:
- population_size=5000
- mutation_rate=0.2
- tournament_size=100
- max_iterations=20000
- selection_method='tournament_selection'
- elitism_size=2500

Since the size of the smallest dominant set for this graph is 9, I manually stopped the execution after the 187th iteration (~3 min) to verify the correctness of the solution. I used the NetworkX library for the verification, which confirmed that the solution is correct.